refactor(configeditor): address review feedback on QWK API screen
- Move sysFieldsQWKAPI into fields_qwkapi.go (fields_system.go was
  already >700 lines against the 300-line guideline).
- Display Token TTL as 24 for any non-positive stored value, matching
  QWKAPIConfig.TokenTTL runtime behavior; test extended.
- Render the sys-menu hotkey for item 10 as '0.' so the displayed key
  matches the input mapping.
